backend/requirements.txt
blacksheep
uvicorn
google-genai
python-dotenv
pydantic-settings
